aphrodite
Module arch

Arch-specific code.

Modules

egatext

Stuff for writing and reading to the EGA text buffer.

interrupts

Provides interrupt-related functions

output

Functions to output to various things

paging

Functions and types related to paging.

ports

Provides utilities for interacting with assembly ports

Constants

DEBUG_PORT

The assembly port number to output debug messages to.

Functions

cpuid

Returns information from the CPUID command in the form (ebx, edx, ecx).

cpuid_extended_functions

Returns whether extended functions are available (more specifically, 0x80000001 or higher)

disable_paging

Disables paging by clearing bit 31 in the cr0 register.

enable_a20

Tries to enable the a20 gate by trying many different methods and seeing what sticks.

enable_a20_ee_port

Tries to enable the a20 gate by reading from port 0xee.

enable_a20_fasta20

Tries to enable the a20 gate via fast a20. Note that this may not work or do something unexpected.

enable_a20_keyboard

Tries to enable the a20 gate via the keyboard controller method.

get_keyboard_data

Gets data from the keyboard.

send_keyboard_cmd

Sends a keyboard command.

send_keyboard_data

Sends data to the keyboard.

test_a20

Returns whether the a20 gate is enabled.

wait_for_keyboard_cmd

Waits for a keyboard command to complete.

wait_for_keyboard_data

Waits for there to be data to read from the keyboard.