aphrodite::arch
Module output

Functions to output to various things

Functions

sdebugb

Outputs a $func_name message &[u8] to the debug serial port.

sdebugbln

Outputs a $func_name message &[u8] and a newline to the debug serial port.

sdebugbnp

Outputs a $func_name message &[u8] to the debug serial port without a prefix.

sdebugbnpln

Outputs a $func_name message &[u8] and a newline to the debug serial port without a prefix.

sdebugs

Outputs a $func_name message &str to the debug serial port.

sdebugsln

Outputs a $func_name message &str and a newline to the debug serial port.

sdebugsnp

Outputs a $func_name message &str to the debug serial port without a prefix.

sdebugsnpln

Outputs a $func_name message &str and a newline to the debug serial port without a prefix.

sdebugu

Outputs a(n) $func_name message u8 to the debug serial port.

sdebugunp

Outputs a(n) $func_name message u8 to the debug serial port without a prefix.

serrorb

Outputs a $func_name message &[u8] to the debug serial port.

serrorbln

Outputs a $func_name message &[u8] and a newline to the debug serial port.

serrorbnp

Outputs a $func_name message &[u8] to the debug serial port without a prefix.

serrorbnpln

Outputs a $func_name message &[u8] and a newline to the debug serial port without a prefix.

serrors

Outputs a $func_name message &str to the debug serial port.

serrorsln

Outputs a $func_name message &str and a newline to the debug serial port.

serrorsnp

Outputs a $func_name message &str to the debug serial port without a prefix.

serrorsnpln

Outputs a $func_name message &str and a newline to the debug serial port without a prefix.

serroru

Outputs a(n) $func_name message u8 to the debug serial port.

serrorunp

Outputs a(n) $func_name message u8 to the debug serial port without a prefix.

sfatalb

Outputs a $func_name message &[u8] to the debug serial port.

sfatalbln

Outputs a $func_name message &[u8] and a newline to the debug serial port.

sfatalbnp

Outputs a $func_name message &[u8] to the debug serial port without a prefix.

sfatalbnpln

Outputs a $func_name message &[u8] and a newline to the debug serial port without a prefix.

sfatals

Outputs a $func_name message &str to the debug serial port.

sfatalsln

Outputs a $func_name message &str and a newline to the debug serial port.

sfatalsnp

Outputs a $func_name message &str to the debug serial port without a prefix.

sfatalsnpln

Outputs a $func_name message &str and a newline to the debug serial port without a prefix.

sfatalu

Outputs a(n) $func_name message u8 to the debug serial port.

sfatalunp

Outputs a(n) $func_name message u8 to the debug serial port without a prefix.

sinfob

Outputs a $func_name message &[u8] to the debug serial port.

sinfobln

Outputs a $func_name message &[u8] and a newline to the debug serial port.

sinfobnp

Outputs a $func_name message &[u8] to the debug serial port without a prefix.

sinfobnpln

Outputs a $func_name message &[u8] and a newline to the debug serial port without a prefix.

sinfos

Outputs a $func_name message &str to the debug serial port.

sinfosln

Outputs a $func_name message &str and a newline to the debug serial port.

sinfosnp

Outputs a $func_name message &str to the debug serial port without a prefix.

sinfosnpln

Outputs a $func_name message &str and a newline to the debug serial port without a prefix.

sinfou

Outputs a(n) $func_name message u8 to the debug serial port.

sinfounp

Outputs a(n) $func_name message u8 to the debug serial port without a prefix.

soutputb

Outputs a $func_name message &[u8] to the debug serial port.

soutputbln

Outputs a $func_name message &[u8] and a newline to the debug serial port.

soutputbnp

Outputs a $func_name message &[u8] to the debug serial port without a prefix.

soutputbnpln

Outputs a $func_name message &[u8] and a newline to the debug serial port without a prefix.

soutputs

Outputs a $func_name message &str to the debug serial port.

soutputsln

Outputs a $func_name message &str and a newline to the debug serial port.

soutputsnp

Outputs a $func_name message &str to the debug serial port without a prefix.

soutputsnpln

Outputs a $func_name message &str and a newline to the debug serial port without a prefix.

soutputu

Outputs a(n) $func_name message u8 to the debug serial port.

soutputunp

Outputs a(n) $func_name message u8 to the debug serial port without a prefix.

swarningb

Outputs a $func_name message &[u8] to the debug serial port.

swarningbln

Outputs a $func_name message &[u8] and a newline to the debug serial port.

swarningbnp

Outputs a $func_name message &[u8] to the debug serial port without a prefix.

swarningbnpln

Outputs a $func_name message &[u8] and a newline to the debug serial port without a prefix.

swarnings

Outputs a $func_name message &str to the debug serial port.

swarningsln

Outputs a $func_name message &str and a newline to the debug serial port.

swarningsnp

Outputs a $func_name message &str to the debug serial port without a prefix.

swarningsnpln

Outputs a $func_name message &str and a newline to the debug serial port without a prefix.

swarningu

Outputs a(n) $func_name message u8 to the debug serial port.

swarningunp

Outputs a(n) $func_name message u8 to the debug serial port without a prefix.