aphrodite
Module display

Types, constants and traits for displaying text. Mostly implemented in arch/.

Constants

COLOR_BLACK

Black-on-black.

COLOR_DEFAULT

Should be whatever color commonly used for status messages. Generally should be white-on-black. Value is one.

Traits

TextDisplay

Some form of display that can be written to with text.

Type Aliases

Color

A type used for color in the functions of TextDisplay.