aphrodite
Module multiboot2

Definitions of structs for multiboot2 information. Mostly used during pre-userspace.

Structs

FramebufferInfo

Information about the framebuffer.

MemoryMap

A full memory map provided by a Multiboot2 bootloader.

MemorySection

One memory section provided by a Multiboot2 bootloader.

Module

A Multiboot2 module. See https://aphrodite-os.github.io/book/bootloader-modules.html.

Multiboot2BootInfo

Boot info collected from provided Tags.

PaletteColorDescriptor

A color descriptor for ColorInfo::Palette.

RawMemoryMap

The raw memory map provided by a Multiboot2 bootloader. This is interpreted into a MemoryMap.

RootTag

The root tag. The official Multiboot2 name is literally the “fixed part” of the tags, so I made a better name.

Tag

Used for Multiboot2 tags. This shouldn’t be used after a crate::boot::BootInfo struct has been initalized, but it still can be used.

Enums

ColorInfo

Information about color, for use in FramebufferInfo.