aphrodite
Module output

Architecture-independent output functions.

Functions

sreset

Resets the position of output to the screen.

tdebugb

Outputs a message &[u8] to the terminal.

tdebugbln

Outputs a message &[u8] and a newline to the terminal.

tdebugbnp

Outputs a message &[u8] to the terminal without a prefix.

tdebugbnpln

Outputs a message &[u8] and a newline to the terminal without a prefix.

tdebugs

Outputs a message &str to the terminal.

tdebugsln

Outputs a message &str and a newline to the terminal.

tdebugsnp

Outputs a message &str to the terminal without a prefix.

tdebugsnpln

Outputs a message &str and a newline to the terminal without a prefix.

tdebugu

Outputs a message u8 to the terminal.

tdebugunp

Outputs a message u8 to the terminal without a prefix.

terrorb

Outputs a message &[u8] to the terminal.

terrorbln

Outputs a message &[u8] and a newline to the terminal.

terrorbnp

Outputs a message &[u8] to the terminal without a prefix.

terrorbnpln

Outputs a message &[u8] and a newline to the terminal without a prefix.

terrors

Outputs a message &str to the terminal.

terrorsln

Outputs a message &str and a newline to the terminal.

terrorsnp

Outputs a message &str to the terminal without a prefix.

terrorsnpln

Outputs a message &str and a newline to the terminal without a prefix.

terroru

Outputs a message u8 to the terminal.

terrorunp

Outputs a message u8 to the terminal without a prefix.

tfatalb

Outputs a message &[u8] to the terminal.

tfatalbln

Outputs a message &[u8] and a newline to the terminal.

tfatalbnp

Outputs a message &[u8] to the terminal without a prefix.

tfatalbnpln

Outputs a message &[u8] and a newline to the terminal without a prefix.

tfatals

Outputs a message &str to the terminal.

tfatalsln

Outputs a message &str and a newline to the terminal.

tfatalsnp

Outputs a message &str to the terminal without a prefix.

tfatalsnpln

Outputs a message &str and a newline to the terminal without a prefix.

tfatalu

Outputs a message u8 to the terminal.

tfatalunp

Outputs a message u8 to the terminal without a prefix.

tinfob

Outputs a message &[u8] to the terminal.

tinfobln

Outputs a message &[u8] and a newline to the terminal.

tinfobnp

Outputs a message &[u8] to the terminal without a prefix.

tinfobnpln

Outputs a message &[u8] and a newline to the terminal without a prefix.

tinfos

Outputs a message &str to the terminal.

tinfosln

Outputs a message &str and a newline to the terminal.

tinfosnp

Outputs a message &str to the terminal without a prefix.

tinfosnpln

Outputs a message &str and a newline to the terminal without a prefix.

tinfou

Outputs a message u8 to the terminal.

tinfounp

Outputs a message u8 to the terminal without a prefix.

toutputb

Outputs a message &[u8] to the terminal.

toutputbln

Outputs a message &[u8] and a newline to the terminal.

toutputbnp

Outputs a message &[u8] to the terminal without a prefix.

toutputbnpln

Outputs a message &[u8] and a newline to the terminal without a prefix.

toutputs

Outputs a message &str to the terminal.

toutputsln

Outputs a message &str and a newline to the terminal.

toutputsnp

Outputs a message &str to the terminal without a prefix.

toutputsnpln

Outputs a message &str and a newline to the terminal without a prefix.

toutputu

Outputs a message u8 to the terminal.

toutputunp

Outputs a message u8 to the terminal without a prefix.

twarningb

Outputs a message &[u8] to the terminal.

twarningbln

Outputs a message &[u8] and a newline to the terminal.

twarningbnp

Outputs a message &[u8] to the terminal without a prefix.

twarningbnpln

Outputs a message &[u8] and a newline to the terminal without a prefix.

twarnings

Outputs a message &str to the terminal.

twarningsln

Outputs a message &str and a newline to the terminal.

twarningsnp

Outputs a message &str to the terminal without a prefix.

twarningsnpln

Outputs a message &str and a newline to the terminal without a prefix.

twarningu

Outputs a message u8 to the terminal.

twarningunp

Outputs a message u8 to the terminal without a prefix.